InstructionDecoder: Added an opcode-to-type cache.
We have instruction-to-opcode mapping, on searching using the opcode, we
have to traverse the map. (The other way is to have reversed maps, much
more memory with particularly little perf improvement.)

This is partially remedied by maintaining a cache for the
reverse-lookup. This partially alleviates the problem since we only
search for any opcode once. This increases the overall performance by
about 8% (using the tests as benchmarks since we have no better means of
benchmarking as of now.)